# src/test/resources/openapi/approved-openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "technical-interview-demo API",
    "description": "Machine-readable contract for the demo application's supported external /api/** surface. Internal-only overview, documentation, OpenAPI publication, and actuator validation paths are intentionally excluded.",
    "version": "APPROVED"
  },
...
  "paths": {
...
    "/api/categories/{id}": {
      "put": {
        "tags": [
          "Categories"
        ],
        "summary": "Update a category",
        "description": "Requires an authenticated session with the ADMIN role and a valid same-site CSRF header mirrored from the readable XSRF-TOKEN cookie.",
        "operationId": "update_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "X-XSRF-TOKEN",
            "in": "header",
            "description": "Same-site CSRF header whose value must match the readable XSRF-TOKEN cookie.",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authenticated session.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated user does not have the ADMIN role.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemResponse"
                }
              }
            }
          },
          "404": {
            "description": "Requested category was not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Categories"
        ],
        "summary": "Delete a category",
        "description": "Requires an authenticated session with the ADMIN role, a valid same-site CSRF header mirrored from the readable XSRF-TOKEN cookie, and a category that is not still assigned to books.",
        "operationId": "delete_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "X-XSRF-TOKEN",
            "in": "header",
            "description": "Same-site CSRF header whose value must match the readable XSRF-TOKEN cookie.",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Missing or invalid authenticated session.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated user does not have the ADMIN role.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemResponse"
                }
              }
            }
          },
          "404": {
            "description": "Requested category was not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemResponse"
                }
              }
            }
          },
          "409": {
            "description": "Category is still assigned to one or more books.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
    "/api/categories": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "List categories",
        "description": "Public endpoint that returns categories ordered by name.",
        "operationId": "findAll_1",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Categories"
        ],
        "summary": "Create a category",
        "description": "Requires an authenticated session with the ADMIN role and a valid same-site CSRF header mirrored from the readable XSRF-TOKEN cookie.",
        "operationId": "create_1",
        "parameters": [
          {
            "name": "X-XSRF-TOKEN",
            "in": "header",
            "description": "Same-site CSRF header whose value must match the readable XSRF-TOKEN cookie.",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authenticated session.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated user does not have the ADMIN role.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CategoryUpdateRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "minLength": 1,
            "type": "string",
            "description": "Unique category name.",
            "example": "Software Architecture"
          }
        },
        "description": "Payload for updating a category."
      },
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Database identifier.",
            "format": "int64",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Unique category name.",
            "example": "Java"
          }
        },
        "description": "Book category."
      },
      "ApiProblemResponse": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Problem title.",
            "example": "Unauthorized"
          },
          "status": {
            "type": "integer",
            "description": "HTTP status code.",
            "format": "int32",
            "example": 401
          },
          "detail": {
            "type": "string",
            "description": "Technical problem detail kept stable for debugging and logs.",
            "example": "Authentication is required to access this resource."
          },
          "messageKey": {
            "type": "string",
            "description": "Stable localization key for the error type.",
            "example": "error.request.unauthorized"
          },
          "message": {
            "type": "string",
            "description": "Localized end-user message resolved from the request language.",
            "example": "You must authenticate before performing this operation."
          },
          "language": {
            "type": "string",
            "description": "Two-letter ISO 639-1 language code actually used for the localized message.",
            "example": "en"
          }
        },
        "description": "Localized ProblemDetail payload returned for API errors."
      },
...
      "CategoryCreateRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "minLength": 1,
            "type": "string",
            "description": "Unique category name.",
            "example": "Software Engineering"
          }
        },
        "description": "Payload for creating a category."
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "description": "Authenticated browser session cookie used by protected operations. It is established through a configured identity provider login path under /api/session/oauth2/authorization/{registrationId} when the optional oauth profile is active. Unsafe browser writes also require the X-XSRF-TOKEN request header mirrored from the readable XSRF-TOKEN cookie.",
        "name": "technical-interview-demo-session",
        "in": "cookie"
      }
    }
  }
}